requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
celery==5.3.4
redis==5.0.1
mysqlclient==2.2.4
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
Pillow==10.2.0
dj-database-url
python-json-logger

# Data Science & ML
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==21.0.0
black==23.12.1
isort==5.13.2
flake8==7.0.0

# Documentation
drf-spectacular==0.27.0